Simplify & improve canvas renderer (#64)
The canvas renderer has had rounding issues forever, if the size of the
QRCode wasn't a multiple of the number of cells being rendered, which
you can't (fairly) predict as a user.

We can make it better. Path2D allows using SVG paths in canvas. We're
already generating a (mostly) optimized single path for the SVG renderer
so this is a quick plug and go. Path2D is pretty widely supported so
this is an easy win.

For the fallback case, I opted to just scale our units so we have the
number of squares to match, eliminating all rounding on our part. We
just draw "1x1" squares which are scaled to match the actual size. This
leaves a small side effect of a single "antialiased" pixel between
squares, but it still looks better.